\name{AchalNLP-package}
\alias{AchalNLP-package}
\alias{AchalNLP}
\docType{package}
\title{
\packageTitle{AchalNLP}
}
\description{
\packageDescription{AchalNLP}
}

\author{
\packageAuthor{AchalNLP}

Maintainer: \packageMaintainer{AchalNLP}
}

\keyword{ package }
\seealso{
~~ \code{\link{sampleTextData}} \code{\link{generateTDM}} \code{\link{predict_Backoff}} ~~
}
