\name{ANLP-package}
\alias{ANLP-package}
\alias{ANLP}
\docType{package}
\title{
\packageTitle{ANLP}
}
\description{
\packageDescription{ANLP}
}

\author{
\packageAuthor{ANLP}

Maintainer: \packageMaintainer{ANLP}
}

\keyword{ package }
\seealso{
~~ \code{\link{sampleTextData}} \code{\link{generateTDM}} \code{\link{predict_Backoff}} ~~
}
