% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANLP.R
\name{buildNgramModel}
\alias{buildNgramModel}
\title{Build N gram model}
\usage{
buildNgramModel(N)
}
\arguments{
\item{N}{size of n-gram model}
}
\value{
function which can be used to build N-gram model
}
\description{
This function is an abstract function used by \code{\link{generateTDM}}
}
\seealso{
\code{\link{NGramTokenizer}} \code{\link{Weka_control}}
}

