% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANLP.R
\name{cleanTextData}
\alias{cleanTextData}
\title{Clean and tokenize string data}
\usage{
cleanTextData(data)
}
\arguments{
\item{data}{Data read by \code{\link{readTextFile}}}
}
\value{
a list having sampled text data
}
\description{
This function applies different cleaning techniques to clean corpus data.
}
\details{
This function removes non english characters, numbers, white spaces, brackets, punctuation. It also handles cases like abbreviation, contraction. It converts entire text to lower case.
}
\seealso{
\code{\link{tm_map}} \code{\link{iconv}} \code{\link{content_transformer}} \code{\link{removeNumbers}} \code{\link{replace_contraction}} \code{\link{replace_abbreviation}} \code{\link{bracketX}} \code{\link{removePunctuation}} \code{\link{tolower}} \code{\link{stripWhitespace}}
}

