% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANLP.R
\name{generateTDM}
\alias{generateTDM}
\title{Generate term document frequency table from corpus}
\usage{
generateTDM(data, N, isTrace = F)
}
\arguments{
\item{data}{It can be text corpus/data cleaned by \code{\link{cleanTextData}}}

\item{N}{size of n-gram model}

\item{isTrace}{for debugging purpose, use this if you want to track time to build model.}
}
\value{
term document matrix for terms having N words
}
\description{
This function builds term documement sparse matrix
}
\details{
This function generates terms with N number of words specified in argument. This can be used in many tasks like information retrival, document similarity etc.
}
\seealso{
\code{\link{TermDocumentMatrix}} \code{\link{buildNgramModel}}
}

