% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANLP.R
\name{predict_Backoff}
\alias{predict_Backoff}
\title{Predict next word using backoff method}
\usage{
predict_Backoff(testline, modelsList, isDebugMode = F)
}
\arguments{
\item{testline}{Line on which we are performing algorithm to predict next word}

\item{modelsList}{List having all Ngram models generated by \code{\link{generateTDM}}}

\item{isDebugMode}{for debugging purpose, this will print out debug statements}
}
\value{
next predicted word
}
\description{
This function predicts next word using back-off algorithm.
}
\details{
This function predicts next word based on previous N number of words using N-gram models generated by \code{\link{generateTDM}}.
}
\seealso{
\code{\link{generateTDM}} \code{\link{TermDocumentMatrix}}
}

