% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANLP.R
\name{readTextFile}
\alias{readTextFile}
\title{Read text files in binary mode}
\usage{
readTextFile(fileName, encoding)
}
\arguments{
\item{fileName}{A path to the text file}

\item{encoding}{Read text data with encoding}
}
\value{
a list having all the text data from file
}
\description{
This function reads text files in the binary mode
}
\examples{
\dontrun{sampleTextData("data/twitter.txt","UTF-8")}
}

