% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANLP.R
\name{sampleTextData}
\alias{sampleTextData}
\title{Sample text data}
\usage{
sampleTextData(data, proportion)
}
\arguments{
\item{data}{Data read by \code{\link{readTextFile}}}

\item{proportion}{Value between 0 to 1 which represents portion of data}
}
\value{
a list having sampled text data
}
\description{
This function reads text files in the binary mode
}
\seealso{
\code{\link{rbinom}}
}

