% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ANLP.R
\docType{data}
\name{twitter.data}
\alias{twitter.data}
\title{Twitter dataset}
\format{An object of class \code{character} of length 109091.}
\usage{
twitter.data
}
\description{
This dataset is having actual tweets. It contains more than 100k tweets.
}
\keyword{datasets}

